[버그] 리렌더링 이슈로 fetch 위치 변경
