+++ frontend/src/components/FriendSearch/FriendSearchHeader.jsx
@@ -38,12 +38,12 @@ const SEARCH_OPTIONS = [
   { name: '강아지 이름', value: 'dogName' },
 ];
 
-function FriendSearchHeader({ initialName = '', initialType = '' }) {
+function FriendSearchHeader({ initialName = '', initialType = '', onSubmit }) {
+  const navigate = useNavigate();
   const [{ searchName }, onChange] = useInputs({
     searchName: initialName,
   });
   const [searchType, setSearchType] = useState(initialType);
-  const naviate = useNavigate();
   const defaultOption = SEARCH_OPTIONS.filter(
     option => option.value === initialType,
   )[0].name;
@@ -59,7 +59,8 @@ function FriendSearchHeader({ initialName = '', initialType = '' }) {
     }
 
     if (event.key === 'Enter') {
-      naviate(
+      onSubmit({ newSearchName: searchName, newSearchType: searchType });
+      navigate(
         `/friend/search?searchType=${searchType}&searchName=${searchName}`,
       );
     }
